+++ atspi-common/src/events/mouse.rs
@@ -1,86 +1,15 @@
 use crate::{
 	error::AtspiError,
-	events::{
-		BusProperties, EventBodyOwned, HasInterfaceName, HasMatchRule, HasRegistryEventString,
-	},
-	Event, EventProperties, EventTypeProperties,
+	events::{BusProperties, EventBodyOwned},
+	Event, EventProperties,
 };
 #[cfg(feature = "zbus")]
 use crate::{
-	events::{
-		EventWrapperMessageConversion, MessageConversion, MessageConversionExt, TryFromMessage,
-	},
+	events::{MessageConversion, MessageConversionExt},
 	ObjectRef,
 };
 use zbus_names::UniqueName;
 use zvariant::ObjectPath;
-
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize, PartialEq, Eq, Hash)]
-pub enum MouseEvents {
-	/// See: [`AbsEvent`].
-	Abs(AbsEvent),
-	/// See: [`RelEvent`].
-	Rel(RelEvent),
-	/// See: [`ButtonEvent`].
-	Button(ButtonEvent),
-}
-
-impl EventTypeProperties for MouseEvents {
-	fn member(&self) -> &'static str {
-		match self {
-			Self::Abs(inner) => inner.member(),
-			Self::Rel(inner) => inner.member(),
-			Self::Button(inner) => inner.member(),
-		}
-	}
-	fn interface(&self) -> &'static str {
-		match self {
-			Self::Abs(inner) => inner.interface(),
-			Self::Rel(inner) => inner.interface(),
-			Self::Button(inner) => inner.interface(),
-		}
-	}
-	fn match_rule(&self) -> &'static str {
-		match self {
-			Self::Abs(inner) => inner.match_rule(),
-			Self::Rel(inner) => inner.match_rule(),
-			Self::Button(inner) => inner.match_rule(),
-		}
-	}
-	fn registry_string(&self) -> &'static str {
-		match self {
-			Self::Abs(inner) => inner.registry_string(),
-			Self::Rel(inner) => inner.registry_string(),
-			Self::Button(inner) => inner.registry_string(),
-		}
-	}
-}
-
-impl EventProperties for MouseEvents {
-	fn path(&self) -> ObjectPath<'_> {
-		match self {
-			Self::Abs(inner) => inner.path(),
-			Self::Rel(inner) => inner.path(),
-			Self::Button(inner) => inner.path(),
-		}
-	}
-	fn sender(&self) -> UniqueName<'_> {
-		match self {
-			Self::Abs(inner) => inner.sender(),
-			Self::Rel(inner) => inner.sender(),
-			Self::Button(inner) => inner.sender(),
-		}
-	}
-}
-
-impl_from_interface_event_enum_for_event!(MouseEvents, Event::Mouse);
-impl_try_from_event_for_user_facing_event_type!(MouseEvents, Event::Mouse);
-
-event_wrapper_test_cases!(MouseEvents, AbsEvent);
-
-impl HasMatchRule for MouseEvents {
-	const MATCH_RULE_STRING: &'static str = "type='signal',interface='org.a11y.atspi.Event.Mouse'";
-}
 
 #[derive(Debug, PartialEq, Clone, serde::Serialize, serde::Deserialize, Eq, Hash, Default)]
 pub struct AbsEvent {
@@ -187,37 +116,8 @@ impl MessageConversion for ButtonEvent {
 		copy.into()
 	}
 }
-impl HasInterfaceName for MouseEvents {
-	const DBUS_INTERFACE: &'static str = "org.a11y.atspi.Event.Mouse";
-}
 
-#[cfg(feature = "zbus")]
-impl EventWrapperMessageConversion for MouseEvents {
-	fn try_from_message_interface_checked(msg: &zbus::Message) -> Result<Self, AtspiError> {
-		let header = msg.header();
-		let member = header.member().ok_or(AtspiError::MissingMember)?;
-		match member.as_str() {
-			AbsEvent::DBUS_MEMBER => Ok(MouseEvents::Abs(AbsEvent::from_message_unchecked(msg)?)),
-			RelEvent::DBUS_MEMBER => Ok(MouseEvents::Rel(RelEvent::from_message_unchecked(msg)?)),
-			ButtonEvent::DBUS_MEMBER => {
-				Ok(MouseEvents::Button(ButtonEvent::from_message_unchecked(msg)?))
-			}
-			_ => Err(AtspiError::MemberMatch("No matching member for Mouse".into())),
-		}
-	}
-}
-
-#[cfg(feature = "zbus")]
-impl TryFrom<&zbus::Message> for MouseEvents {
-	type Error = AtspiError;
-	fn try_from(msg: &zbus::Message) -> Result<Self, Self::Error> {
-		Self::try_from_message(msg)
-	}
-}
-
-impl_from_user_facing_event_for_interface_event_enum!(AbsEvent, MouseEvents, MouseEvents::Abs);
 impl_from_user_facing_type_for_event_enum!(AbsEvent, Event::Mouse);
-impl_try_from_event_for_user_facing_type!(AbsEvent, MouseEvents::Abs, Event::Mouse);
 
 event_test_cases!(AbsEvent);
 impl_to_dbus_message!(AbsEvent);
@@ -235,9 +135,7 @@ impl From<AbsEvent> for EventBodyOwned {
 	}
 }
 
-impl_from_user_facing_event_for_interface_event_enum!(RelEvent, MouseEvents, MouseEvents::Rel);
 impl_from_user_facing_type_for_event_enum!(RelEvent, Event::Mouse);
-impl_try_from_event_for_user_facing_type!(RelEvent, MouseEvents::Rel, Event::Mouse);
 event_test_cases!(RelEvent);
 impl_to_dbus_message!(RelEvent);
 impl_from_dbus_message!(RelEvent);
@@ -254,13 +152,7 @@ impl From<RelEvent> for EventBodyOwned {
 	}
 }
 
-impl_from_user_facing_event_for_interface_event_enum!(
-	ButtonEvent,
-	MouseEvents,
-	MouseEvents::Button
-);
 impl_from_user_facing_type_for_event_enum!(ButtonEvent, Event::Mouse);
-impl_try_from_event_for_user_facing_type!(ButtonEvent, MouseEvents::Button, Event::Mouse);
 event_test_cases!(ButtonEvent);
 impl_to_dbus_message!(ButtonEvent);
 impl_from_dbus_message!(ButtonEvent);
@@ -275,8 +167,4 @@ impl From<ButtonEvent> for EventBodyOwned {
 			any_data: u8::default().into(),
 		}
 	}
-}
-
-impl HasRegistryEventString for MouseEvents {
-	const REGISTRY_EVENT_STRING: &'static str = "Mouse:";
 }
